+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added collapse button to library sidebar
 * Dictionary search bar now uses fuzzy search (can accomodate some typos)
 * Patched image renderer. It's faster now.
+* The default channel and the unread marker are now localized
 
 ### Fixes